feat(analysisType): Refactor
