+++ cg/constants/tb.py
@@ -20,3 +20,10 @@ class AnalysisTypes(StrEnum):
     WES: str = SeqLibraryPrepCategory.WHOLE_EXOME_SEQUENCING
     WGS: str = SeqLibraryPrepCategory.WHOLE_GENOME_SEQUENCING
     WTS: str = SeqLibraryPrepCategory.WHOLE_TRANSCRIPTOME_SEQUENCING
+
+
+class AnalysisType(StrEnum):
+    TGS: str = "tgs"
+    WES: str = "wes"
+    WGS: str = "wgs"
+    WTS: str = "wts"
