+++ cg/meta/delivery_report/balsamic.py
@@ -17,8 +17,8 @@ from cg.constants import (
     REQUIRED_SAMPLE_TIMESTAMP_FIELDS,
     Workflow,
 )
-from cg.constants.constants import AnalysisType
 from cg.constants.scout import ScoutUploadKey
+from cg.constants.tb import AnalysisType
 from cg.meta.delivery_report.data_validators import get_million_read_pairs
 from cg.meta.delivery_report.delivery_report_api import DeliveryReportAPI
 from cg.meta.workflow.balsamic import BalsamicAnalysisAPI
@@ -52,9 +52,7 @@ class BalsamicDeliveryReportAPI(DeliveryReportAPI):
         passed_initial_qc: bool | None = self.lims_api.has_sample_passed_initial_qc(
             sample.internal_id
         )
-        if AnalysisType.WHOLE_GENOME_SEQUENCING in self.analysis_api.get_data_analysis_type(
-            case.internal_id
-        ):
+        if AnalysisType.WGS in self.analysis_api.get_data_analysis_type(case.internal_id):
             return self.get_wgs_metadata(
                 million_read_pairs=million_read_pairs,
                 passed_initial_qc=passed_initial_qc,
